refactor: Participant Management 서비스 분할
